test(spec): split criteria anchor/region-scoping tests into their own file
shared/tests/test_layer_coverage_criteria.py crossed its 300-line bloat
baseline (319 lines) after the Stage-3 doubt-review regression tests were
added. Per the bloat-policy invariant (find redundancy/split, never raise
the baseline cap), split the two newest tests - the nested-heading anchor
case and the level-3 Acceptance Criteria heading scoping case, both about
WHERE criteria_digests looks rather than what counts as a criterion once
it's looking - into a new file,
shared/tests/test_layer_coverage_criteria_anchoring.py, the same seam
test_spec_checks_shipped_form.py was already split from test_spec_checks.py
along earlier in this run.

test_layer_coverage_criteria.py: 270 lines (was 319). Both files re-run
clean (21 passed total, same count as before the split), ruff clean, the
anti-ratchet hook clean against staged content.

Run-ID: iterate-2026-08-25-r0-spec-reader-shipped-shape
