+++ .shipwright/triage.jsonl
@@ -1915,3 +1915,4 @@
 {"event":"append","id":"trg-b85ebe2e","ts":"2026-08-25T11:21:20.056872Z","originalTs":"2026-08-25T11:21:20.056872Z","source":"iterate-req3-04-ac-identity-mono-R0","severity":"low","kind":"improvement","title":"project/adopt emit no per-FR description block — decide whether they should","detail":"R0 (campaign req3-04-ac-identity-mono) converged the three FR-criteria readers on the shipped shape: a heading followed by bare (E)/checkbox bullets, description read from the requirements table row instead of a **Description:** label. That is a deliberate choice by /shipwright-project and /shipwright-adopt (FR-01.02, FR-01.13) — no per-FR description block is ever emitted, only the table cell. Whether that is the right call long-term (a table cell is terse; a labelled prose block would carry more context) is an open product question R0 did not decide — it only made the readers stop mis-reporting the current shape. Filed per R0's Acceptance Criteria so the question is not lost.","evidencePath":null,"runId":null,"commit":null,"dedupKey":null,"launchPayload":null,"frId":"FR-01.02","suiteId":null,"eventId":null,"status":"triage","suggestedPriority":"P3","suggestedDomain":"engineering"}
 {"event":"append","id":"trg-892387fe","ts":"2026-08-25T12:43:00.854500Z","originalTs":"2026-08-25T12:43:00.854500Z","source":"iterate-req3-04-ac-identity-mono-R0","severity":"low","kind":"improvement","title":"Campaign sub-iterate run_ids must not embed an uppercase display id (RUN_ID_STRICT + 3 sibling regexes are lowercase-only)","detail":"R0 (campaign req3-04-ac-identity-mono) was minted as iterate-2026-08-25-R0-spec-reader-shipped-shape and failed append_iterate_entry.py's RUN_ID_STRICT closed at F5c (lowercase-only; no strict=False path for a new entry). Same lowercase-only shape is duplicated in lib/agent_doc_shape.py, tools/verifiers/ci_supplychain.py and scripts/accepted_risks.py. Worked around this run by renaming the run_id to lowercase end-to-end (F3/F4/F5/F5b/F5c); see the ADR's Reflection section. Whoever mints campaign sub-iterate run_ids should pre-lowercase the embedded id (e.g. r0, not R0) so future sub-iterates do not hit the same wall.","evidencePath":null,"runId":"iterate-2026-08-25-r0-spec-reader-shipped-shape","commit":null,"dedupKey":null,"launchPayload":"/shipwright-iterate <id>","frId":null,"suiteId":null,"eventId":null,"status":"triage","suggestedPriority":"P3","suggestedDomain":"engineering"}
 {"event":"append","id":"trg-968e4d87","ts":"2026-08-25T13:51:17.135977Z","originalTs":"2026-08-25T13:51:17.135977Z","source":"code-review","severity":"low","kind":"improvement","title":"FR-criteria convergence: two newly-widened parsing rules are unlisted/untested","detail":"PR #648 (run_id iterate-2026-08-25-r0-spec-reader-shipped-shape) unified 3 FR-criteria readers onto shared/scripts/lib/fr_criteria.py. Stage-2 code review flagged two areas as safe-to-defer, not fixed in that PR: (1) the cross-layer fold gate's iter_anchored_blocks widened block-termination surface vs the old narrower parser has two rules with no dedicated test or doc mention -- a same-or-lower non-FR heading now ends a block, and a criterion line starting with **FR-XX.YY now truncates the block; (2) group_i_criteria (I6)'s own bullet semantics silently widened too (ordered lists, placeholder+continuation lines) with no explicit pin. Neither is a known bug -- both are follow-up test/doc coverage.","evidencePath":null,"runId":"iterate-2026-08-25-r0-spec-reader-shipped-shape","commit":null,"dedupKey":null,"launchPayload":null,"frId":null,"suiteId":null,"eventId":null,"status":"triage","suggestedPriority":"P3","suggestedDomain":"engineering"}
+{"event":"append","id":"trg-467b7b2f","ts":"2026-08-25T14:49:54.885651Z","originalTs":"2026-08-25T14:49:54.885651Z","source":"doubt-reviewer","severity":"low","kind":"improvement","title":"fr_criteria: table-row exemption and I6 convergence test have two narrow follow-up gaps","detail":"From R0's (campaign req3-04-ac-identity-mono) doubt-review round 1, low severity, deferred as non-blocking: (1) the FR-table description exemption can fire on a table with no real Description column, falling back to the Name cell; (2) the three-way convergence test's I6 arm reaches the shared parser through the plugin module's attribute rather than an I6-owned entry point, so it proves reachability more than I6's own contract. See trg-968e4d87 for the sibling round-1 deferrals (cross-layer block-termination, I6 bullet-semantics widening) from the same run.","evidencePath":null,"runId":null,"commit":null,"dedupKey":null,"launchPayload":null,"frId":null,"suiteId":null,"eventId":null,"status":"triage","suggestedPriority":"P3","suggestedDomain":"engineering"}
